+++ lib/EnzymeCore/Project.toml
@@ -1,0 +1,11 @@
+name = "EnzymeCore"
+uuid = "f151be2c-9106-41f4-ab19-57ee4f262869"
+authors = ["William Moses <wmoses@mit.edu>", "Valentin Churavy <vchuravy@mit.edu>"]
+version = "0.1.0"
+
+[deps]
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+
+[compat]
+Adapt = "3.3"
+julia = "1.6"
